libplacebo: use -Db_staticpic instead of --b_staticpic
The documentation states that these can be used interchangeably, but
building it just now it gave me the following error:

meson: error: unrecognized arguments: --b_staticpic=false

So apparently the documentation is wrong or perhaps the --b_staticpic
syntax only works on very new versions of meson.

Probably caused by me not testing the final version of of PR#2 properly.